pyproject.toml
# pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wash-alpha"
version = "2.0.0"
description = "Cross-account wash sale detection — Schwab CSV, options, ETFs (v2 simplified)"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "sqlmodel>=0.0.16",
    "typer[all]>=0.12",
    "loguru>=0.7",
    "pyyaml>=6.0",
]

[project.scripts]
net-alpha = "net_alpha.cli.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "factory-boy>=3.3",
    "ruff>=0.6",
]

[tool.hatch.build.targets.wheel]
packages = ["src/net_alpha"]

[tool.ruff]
target-version = "py311"
src = ["src"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "master"
build_command = ""
tag_format = "v{version}"
commit_message = "chore: release v{version}"
